feat: :sparkles: Add team abbreviations
Create team abbreviations on each team to allow for retrieving a team and using their abbreviation.
Prior, abbreviations could be found in `nicknames` and still can be, but this is for use cases where a team is found and more readily available information is provided to the user, such as the abbreviation along with the other data.